chore: move Sertificate types to zod scheme
